Created/tested postman scripts, fix ownership bug
Created and tested postman scripts using postman app in "survey"
collection.  Exported survey collection and provided instructions in
README.md on how to import and execute postman scripts.≠

Update did not work because of commented out code in models/survey.js.
Code was commented out because we thought we were going to defer
ownership, but we decided it was easier to include ownership.

Co-authored-by: Tk <tukrong@gmail.com>