Created/tested postman scripts, fix ownership bug
RAN  INTO REBASE ISSUES - see bottom

Created and tested postman scripts using postman app in "survey"
collection.  Exported survey collection and provided instructions in
README.md on how to import and execute postman scripts.≠

Update did not work because of commented out code in models/survey.js.
Code was commented out because we thought we were going to defer
ownership, but we decided it was easier to include ownership.

Merge branch 'versionone' of
github.com:wat-the-duck/pollaris-surveys-api into versionone

When I did git rebase, two files seemed to already be in dev,
don't know how to get it, but we retested after rebasing.  When tried to
do a push, told us we needed to pull from version one.

Co-authored-by: Tk <tukrong@gmail.com>